Add files via upload
OTP Verification is the process of verifying a user by sending a 
unique password so that the user can be verified before 
completing a registration or payment process. Most of the time, 
we get an OTP when we make an online payment, or when we 
forget our password, or when creating an account on any online 
platform. Thus, the sole purpose of an OTP is to verify the 
identity of a user by sending a unique password.